% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs1.welch.adj.R
\name{rhs1.welch.adj}
\alias{rhs1.welch.adj}
\title{rhs1.welch.adj}
\usage{
rhs1.welch.adj(z, rho, f1, f2, alpha, const)
}
\arguments{
\item{alpha}{The nominal coverage probability is 1-alpha}

\item{const}{}
}
\value{

}
\description{
Written by P.Kabaila for the following paper:
Kabaila, P. (2005). Assessment of a preliminary
F-test solution to the Behrens-Fisher problem.
Communications in Statistics - Theory and Methods,
34, 291-302.
This function may be used freely, provided the
user makes reference to this paper.
This function calculates
 c C_{\alpha/2}(r)
 expressed as a function of z.
The notation used in this function is the same as
in this paper.
}
