% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.to.r.R
\name{z.to.r}
\alias{z.to.r}
\title{z.to.r}
\usage{
z.to.r(z, rho, f1, f2)
}
\arguments{
\item{z}{This is the parameter z}

\item{rho}{This is the parameter rho}

\item{f1}{First degrees of freedom}

\item{f2}{Second degrees of freedom}
}
\description{
Written by P.Kabaila for the following paper:
Kabaila, P. (2005). Assessment of a preliminary
F-test solution to the Behrens-Fisher problem.
Communications in Statistics - Theory and Methods,
34, 291-302.
This function may be used freely, provided the
user makes reference to this paper.
This function converts z to r.
The notation used in this function is the same as
in this paper.
}
